Move only: Move specialized impl block into a module
There are more specialized key types coming
